사용자 인증 모듈 경로 변경 및 관련 참조 업데이트
인증 모듈의 경로를 'applications/auth'로 변경하고, 관련된 모든 참조를 업데이트했습니다.

주요 변경사항:
1. `auth` 모듈의 경로를 `applications/auth`로 변경
2. `app.controller.ts` 및 `app.module.ts`에서 `auth` 모듈 경로 수정
3. `auth` 모듈 내의 모든 파일 경로 및 참조 수정
4. `users`, `dogs`, `journals`, `statistics` 등 다른 모듈에서 `auth` 모듈 참조 경로 수정
