+++ backend/server/src/applications/auth/auth.module.ts
@@ -3,13 +3,13 @@ import { ConfigService } from '@nestjs/config';
 import { APP_GUARD } from '@nestjs/core';
 import { JwtModule } from '@nestjs/jwt';
 
+import { DogsModule } from 'dogs/dogs.module';
+
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { AuthGuard } from './guards/auth.guard';
 import { OauthModule } from './oauth/oauth.module';
 import { TokenService } from './token/token.service';
-
-import { DogsModule } from '../dogs/dogs.module';
 
 @Module({
     imports: [
