+++ backend/server/src/applications/auth/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './oauth-data.decorator';
+export * from './public.decorator';
