+++ backend/server/src/applications/auth/dtos/index.ts
@@ -1,0 +1,2 @@
+export * from './oauth-authorize.dto';
+export * from './oauth.dto';
