+++ backend/server/src/applications/auth/guards/index.ts
@@ -1,0 +1,3 @@
+export * from './auth.guard';
+export * from './oauth-data.guard';
+export * from './refresh-token.guard';
