+++ backend/server/src/applications/auth/index.ts
@@ -1,0 +1,11 @@
+export * from './decorators';
+export * from './dtos';
+export * from './guards';
+export * from './interceptors';
+export * from './oauth';
+export * from './token';
+export * from './types';
+export * from './auth.controller';
+export * from './auth.module';
+export * from './auth.service.spec';
+export * from './auth.service';
