+++ backend/server/src/applications/auth/interceptors/cookie.interceptor.ts
