+++ backend/server/src/applications/auth/interceptors/index.ts
@@ -1,0 +1,1 @@
+export * from './cookie.interceptor';
