+++ backend/server/src/applications/auth/oauth/oauth.module.ts
@@ -3,12 +3,12 @@ import { Module, Type } from '@nestjs/common';
 
 import { ConfigService } from '@nestjs/config';
 
+import { WinstonLoggerService } from 'shared/logger';
+
 import { GoogleService } from './google.service';
 import { KakaoService } from './kakao.service';
 import { NaverService } from './naver.service';
 import { OauthService } from './oauth.service.base';
-
-import { WinstonLoggerService } from '../../shared/logger/winstonLogger.service';
 
 export const OAUTH_REGISTRY = new Map<string, Type<OauthService>>([
     ['google', GoogleService],
