+++ backend/server/src/applications/auth/token/index.ts
@@ -1,0 +1,1 @@
+export * from './token.service';
