+++ backend/server/src/applications/auth/token/token.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtService } from '@nestjs/jwt';
 
-import { parse } from '../../utils/ms.util';
+import { parse } from 'utils/ms.util';
+
 import { OauthProvider } from '../types/oauth-provider.type';
 
 export interface AccessTokenPayload {
