+++ backend/server/src/applications/auth/types/auth-data.type.ts
