+++ backend/server/src/applications/auth/types/index.ts
@@ -1,0 +1,4 @@
+export * from './auth-data.type';
+export * from './oauth-authorize-data.type';
+export * from './oauth-data.type';
+export * from './oauth-provider.type';
