+++ backend/server/src/applications/auth/types/oauth-data.type.ts
