+++ backend/server/src/applications/auth/types/oauth-provider.type.ts
