+++ backend/server/src/journals/journals.controller.ts
@@ -14,13 +14,14 @@ import {
     ValidationPipe,
 } from '@nestjs/common';
 
+import { AccessTokenPayload } from 'applications/auth/token/token.service';
+
 import { CreateJournalDto } from './dtos/create-journal.dto';
 import { UpdateJournalDto } from './dtos/update-journal.dto';
 import { AuthJournalGuard } from './guards/auth-journal.guard';
 import { JournalsService } from './journals.service';
 import { JournalListResponse } from './types/journal.types';
 
-import { AccessTokenPayload } from '../auth/token/token.service';
 import { AuthDogGuard } from '../dogs/guards/auth-dog.guard';
 import { DateValidationPipe } from '../statistics/pipes/date-validation.pipe';
