+++ backend/server/src/statistics/statistics.controller.ts
@@ -1,12 +1,13 @@
 import { Controller, Get, Param, ParseIntPipe, Query, UseGuards } from '@nestjs/common';
 
+import { AccessTokenPayload } from 'applications/auth/token/token.service';
+
 import { DateValidationPipe } from './pipes/date-validation.pipe';
 
 import { Period, PeriodValidationPipe } from './pipes/period-validation.pipe';
 
 import { StatisticsService } from './statistics.service';
 
-import { AccessTokenPayload } from '../auth/token/token.service';
 import { AuthDogGuard } from '../dogs/guards/auth-dog.guard';
 import { User } from '../users/decorators/user.decorator';
